Move zipupload sources to matching package directory
