+++ src/main/java/com/crm/zipupload/entity/InspectionPhoto.java
@@ -1,4 +1,4 @@
-package com.crm.springsecurity.entity;
+package com.crm.zipupload.entity;
 
 import java.time.OffsetDateTime;
